rename query file
