--- app/graphql/queries/company.rb
@@ -1,14 +1,0 @@
-module Queries
-  class Company < Queries::BaseQuery
-    argument :id, ID, required: true
-    type Types::CompanyType, null: true
-
-    def resolve(id:)
-      company_mock = {
-        id: 1,
-        name: 'dev tech stack',
-        default_image_path: ''
-      }
-    end
-  end
-end
