+++ app/graphql/types/query_type.rb
@@ -8,8 +8,8 @@ module Types
     # They will be entry points for queries on your schema.
 
     # field :company, resolver: Queries::Company
-    field :company, resolver: Queries::Company
-    field :stack, resolver: Queries::Stack
-    field :user, resolver: Queries::User::User
+    field :get_company, resolver: Queries::Company::GetCompany
+    field :get_stack, resolver: Queries::Stack::GetStack
+    field :get_user, resolver: Queries::User::GetUser
   end
 end
